requirements.txt
beautifulsoup4==4.5.3
graphitesender==0.11.2
pause==0.2
